requirements.txt
django
pillow
torch
torchvision
boto3
requests
